refac: refatora testes

+++ src/controllers/usersController.ts
@@ -1,5 +1,6 @@
 import { Request, Response } from 'express';
 import * as userService from '../services/userService'
+import { CustomError } from '../Error/CustomError';
 
 const list = async (req: Request, res: Response) => {
   
@@ -23,13 +24,18 @@ const edit = async (req: Request, res: Response) => {
   const { id } = req.params
   const { name, email } = req.params
 
-  const editedUser = await userService.edit(parseInt(id), name, email)
+  const editedUser = await userService.edit(Number(id), name, email)
   return res.status(200).json(editedUser)  
 
 }
 
 const remove = async (req: Request, res: Response) => {
   const { id } = req.params
+  const user = res.locals.user
+
+  if(user.id !== Number(id)){
+    throw new CustomError('You cannot delete another user.', 401)
+  }
   await userService.remove(parseInt(id))
   return res.status(204).json()
